+++ carbon/include/flipper/spi.h
@@ -12,7 +12,7 @@ uint8_t spi_ready(void);
 void spi_put(uint8_t byte);
 void spi_end(void);
 uint8_t spi_get(void);
-int spi_write(void *src, size_t length);
-int spi_read(void *dst, size_t length);
+int spi_write(void *src, uint32_t length);
+int spi_read(void *dst, uint32_t length);
 
 #endif
